refactor(ui): Extract route components to satisfy react-refresh
The updated eslint-plugin-react-refresh flags files that export both
components and non-component values, since Fast Refresh only works when
a module exports components exclusively.

- Move route UI components into src/ui/components/ and read route data
  via getRouteApi(...) to keep route files exporting only their Route
  definition (RootLayout, LandingPage, ProjectVersionsOverview,
  DocumentationComponent)
- Split IFrameScrollContext into a context/hook module and a separate
  IFrameScrollProvider component
- Remove the redundant default export in $projectName/index.tsx
- Fix missing list "key" on the version chips/badges in
  ProjectVersionsOverview (key was on the inner Chip instead of the
  element returned by map)
